feat(product): init product service
The product service has been init with everything needed for it.
It should be relatively easy to add stuff to it now.
This also adds a gitignore that should cover all the languages that
we are going to work with. There may still be a few things that are
not ignored right now that we may ignore later.